+++ tasks/Dotnet-Matrix.Tasks.ps1
@@ -1,8 +1,13 @@
+
+if ( -Not $PsBuildTaskBranch ) {
+    $PsBuildTaskBranch = 'main'
+}
+
 #region InvokeBuild
 
 task UpdateBuildTasks {
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/DotNet/Build.Tasks.ps1' `
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/DotNet/Build.Tasks.ps1" `
         -OutFile "$PSScriptRoot\Build.Tasks.ps1"
 }
 
@@ -13,7 +18,7 @@ task UpdateValidationWorkflow {
     [System.IO.FileInfo] $file = "$PSScriptRoot/../.github/workflows/build-validation.yml"
     New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/build-validation-matrix.yml' `
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/GitHub/build-validation-matrix.yml" `
         -OutFile $file
 }
 
@@ -22,7 +27,7 @@ task UpdatePreReleaseWorkflow {
     [System.IO.FileInfo] $file = "$PSScriptRoot\..\.github\workflows\pre-release.yml"
     New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/pre-release-windows.yml' |
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/GitHub/pre-release-windows.yml" |
     ForEach-Object { $_ -replace 'MyModuleName', $ModuleName } |
     Out-File $file -NoNewline
 }
@@ -32,7 +37,7 @@ task UpdateReleaseWorkflow {
     [System.IO.FileInfo] $file = "$PSScriptRoot\..\.github\workflows\release.yml"
     New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/release-windows.yml' |
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/GitHub/release-windows.yml" |
     ForEach-Object { $_ -replace 'MyModuleName', $ModuleName } |
     Out-File $file -NoNewline
 }
@@ -43,10 +48,10 @@ task UpdateReleaseWorkflow {
 task UpdateIndexPage {
     New-Item -Type Directory "$PSScriptRoot\..\docs" -ErrorAction SilentlyContinue
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/docs/index.md' `
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/docs/index.md" `
         -OutFile "$PSScriptRoot\..\docs\index.md"
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/docs/_config.yml' `
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/docs/_config.yml" `
         -OutFile "$PSScriptRoot\..\docs\_config.yml"
 }
 
@@ -58,7 +63,7 @@ task UpdateDependabotConfig {
     [System.IO.FileInfo] $file = "$PSScriptRoot\..\.github\dependabot.yml"
     New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/dependabot/dependabot.yml' |
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/dependabot/dependabot.yml" |
     ForEach-Object { $_ -replace 'MyModuleName', $ModuleName } |
     Out-File $file -NoNewline
 }
@@ -70,7 +75,15 @@ task UpdateVsCodeTasks {
     [System.IO.FileInfo] $file = "$PSScriptRoot\..\.vscode\tasks.json"
     New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/VsCode/tasks.json' `
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/VsCode/tasks.json" `
+        -OutFile $file
+}
+
+task UpdateVsCodeLaunch {
+    [System.IO.FileInfo] $file = "$PSScriptRoot\..\.vscode\launch.json"
+    New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
+    Invoke-WebRequest `
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/VsCode/launch.json" `
         -OutFile $file
 }
 
@@ -79,10 +92,10 @@ task UpdateVsCodeTasks {
 
 task UpdatePsBuildTasksTasks {
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/tasks/Dotnet-Matrix.Tasks.ps1' `
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/tasks/Dotnet-Matrix.Tasks.ps1" `
         -OutFile "$PSScriptRoot\PsBuild.Tasks.ps1"
 }
 
 #endregion
 
-task UpdatePsBuildTasks -Jobs UpdateBuildTasks, UpdateValidationWorkflow, UpdatePreReleaseWorkflow, UpdateReleaseWorkflow, UpdateIndexPage, UpdateDependabotConfig, UpdateVsCodeTasks, UpdatePsBuildTasksTasks
+task UpdatePsBuildTasks -Jobs UpdateBuildTasks, UpdateValidationWorkflow, UpdatePreReleaseWorkflow, UpdateReleaseWorkflow, UpdateIndexPage, UpdateDependabotConfig, UpdateVsCodeTasks, UpdateVsCodeLaunch, UpdatePsBuildTasksTasks