+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Branch config for Dotnet matrix.
+- Launch config for .Net debugging.
+
 ## [1.4.0] - 2022-08-07
 
 ### Added